Fix time reporting bug
