--- src/utils/timing.h
@@ -1,46 +1,0 @@
-// The MIT License (MIT)
-//
-// Copyright (c) 2021 NVIDIA CORPORATION
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-#include <chrono>
-#include <cstdint>
-
-namespace triton { namespace backend { namespace dali {
-
-inline int64_t capture_time() {
-  auto time = std::chrono::high_resolution_clock::now().time_since_epoch();
-  return std::chrono::duration_cast<std::chrono::nanoseconds>(time).count();
-}
-
-struct TimeInterval {
-  int64_t start = 0;
-  int64_t end = 0;
-};
-
-inline void start_timer_ns(TimeInterval &interval) {
-  interval.start = capture_time();
-}
-
-inline void end_timer_ns(TimeInterval &interval) {
-  interval.end = capture_time();
-}
-
-}}}  // namespace triton::backend::dali
